refactor(cli): 拆分 ts_bridge Bridge DTO 与 CLI 公开类型
- 内部 Bridge 序列化类型迁入 types/bridge.rs
- CliExtension/CliMarketplace 公开类型迁入 types/cli_public.rs
- mod.rs 统一 pub use，外部引用路径不变
