+++ apps/cli/src/ts_bridge/types/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod bridge;
+pub mod cli_public;
+
+pub use cli_public::*;
